requirements.txt
# Data Processing
pandas==2.1.3
numpy==1.24.3

# Database
psycopg2-binary==2.9.9

# API & Web
requests==2.31.0
python-dotenv==1.0.0
PyYAML==6.0.1

# Visualization
streamlit==1.28.1
plotly==5.17.0

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.0

# Code Quality
pre-commit==3.5.0
